requirements.txt
cffi==1.14.6
clipboard==0.0.4
cryptography==3.4.8
ffmpeg-python==0.2.0
future==0.18.2
importlib-metadata==4.8.1
jeepney==0.7.1
keyring==23.2.1
lz4==3.1.3
pbkdf2==1.3
pyaes==1.6.1
pycparser==2.20
pycryptodome==3.10.1
pyperclip==1.8.2
pytube==11.0.1
SecretStorage==3.3.1
typing-extensions==3.10.0.2
zipp==3.5.0
